requirements.txt
discord.py #discord
asyncpg #banco
python-dotenv #esconder token
